Check if the output already exists before running a command.

+++ scripts/merge_fastqs.sh
@@ -6,6 +6,12 @@
 #
 # The directory containing the samples is indicated by the first argument ($1).
 
+if [ -e $2/$3.fastq.gz ]
+then
+	echo "Merged already done."
+	exit 0
+fi
+
 mkdir -p out/merged
 
 	cat $1/$3* > $2/$3.fastq.gz
